+++ bottery/messenger/__init__.py
@@ -1,0 +1,5 @@
+from bottery.messenger.api import MessengerAPI  # noqa
+from bottery.messenger.engine import MessengerEngine
+
+
+engine = MessengerEngine