+++ bottery/platform/telegram.py
@@ -3,8 +3,8 @@
 import warnings
 from bottery.exceptions import BotteryDeprecationWarning
 
-warnings.warn('Module `bottery.platform` is deprecated, '
-             'use `bottery.platforms` instead',
+warnings.warn('Module `bottery.platform.telegram` is deprecated, '
+             'use `bottery.telegram` instead',
              BotteryDeprecationWarning, stacklevel=2)
 
-from bottery.platforms import *
+from bottery.telegram import *