+++ bottery/telegram/engine.py
@@ -3,9 +3,9 @@ import logging
 
 from aiohttp import web
 
-from bottery import platform
 from bottery.message import Message
-from bottery.platform.telegram import TelegramAPI
+from bottery.platforms import BaseEngine
+from bottery.telegram import TelegramAPI
 
 logger = logging.getLogger('bottery.telegram')
 
@@ -56,7 +56,7 @@ class TelegramChat:
         return s.format(u=self)
 
 
-class TelegramEngine(platform.BaseEngine):
+class TelegramEngine(BaseEngine):
     platform = 'telegram'
 
     def __init__(self, *args, **kwargs):