+++ CHANGELOG.md
@@ -5,6 +5,7 @@ _Changes to be released_
 #### Enhancements
 
 - Handler instances returns the selected view ([#161](https://github.com/rougeth/bottery/issues/161))
+- Move Telegram and Messenger modules to outside `platform` module. Also rename `platform` module to `platforms` (due to Python builtin `platform` module) ([#161](https://github.com/rougeth/bottery/issues/165))
 
 #### Bug fixes